Errors in dtogether and some more commands
Hello, I found out that you can't play some games like Betrayal, Doodle-Crew and some more games. It was not in the documentation or somewhere. Plus it never functioned. Always there was the attribute error, that im not in a voice channel.

Then the 8-ball command, you could do that much better like in a embed with a icon, but i gotcha. Could you please add like only the 8-ball command if i am not right with the dtogether command